Create a RailsIntegrator that will handle proxying commands between different versions of rails and how they handle mysql/base sql runners.  Implement this in the current version of the adapter
