+++ lib/departure.rb
@@ -15,17 +15,15 @@ require 'departure/errors'
 require 'departure/command'
 require 'departure/connection_base'
 require 'departure/migration'
+require 'departure/rails_integrator'
 
 require 'departure/railtie' if defined?(Rails)
 
 # We need the OS not to buffer the IO to see pt-osc's output while migrating
 $stdout.sync = true
 
-ActiveSupport.on_load(:active_record) do
-  ActiveRecord::Migration.class_eval do
-    include Departure::Migration
-  end
-end
+
+Departure::RailsIntegrator.for_current.register_integrations
 
 module Departure
   class << self
