+++ spec/spec_helper.rb
@@ -1,5 +1,5 @@
 require 'simplecov'
-SimpleCov.start
+# SimpleCov.start
 
 ENV['RAILS_ENV'] ||= 'development'
 
@@ -18,6 +18,8 @@ require 'support/matchers/have_index'
 require 'support/matchers/have_foreign_key_on'
 require 'support/shared_examples/column_definition_method'
 require 'support/table_methods'
+
+Departure::RailsIntegrator.for_current.register_integrations
 
 db_config = Configuration.new
